refactor: change value type from decimal to double in CreateProductDto and update repository save methods for consistency

+++ src/application/StockManager.Application/Dtos/requests/CreateProductDto.cs
@@ -5,7 +5,7 @@ namespace application.StockManager.Application.Dtos.resquests
     public class CreateProductDto
     {
         public string name { get; set; }
-        public decimal value { get; set; }
+        public double value { get; set; }
         public double amount { get; set; }
         public string? urlImage { get; set; }
         public double discount { get; set; }
